API/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.25.0
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
redis==5.3.0
joblib==1.3.2
pydantic==2.6.1
scikit-learn==1.3.3
python-dotenv==1.1.0
